requirements.txt
tensorflow==2.14.0
tensorflow-addons
tqdm
pillow
numpy